# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Tecnica Jeferson Mercedes",
        "description": "Documentación de la API",
        "contact": {
            "email": "jeferson.md0212@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener todas las divisas",
                "operationId": "73e0a13248f9f6aea618716227eab68a",
                "responses": {
                    "200": {
                        "description": "Lista de divisas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Dólar"
                                            },
                                            "symbol": {
                                                "type": "string",
                                                "example": "$"
                                            },
                                            "exchange_rate": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Crear una nueva divisa",
                "operationId": "57e1c3f26ae9c82bd5996ac5bb738c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Euro"
                                    },
                                    "symbol": {
                                        "type": "string",
                                        "example": "€"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Divisa creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/currencies/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener una divisa por ID",
                "operationId": "23440d4caad0c601f680b655f9c907f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Divisa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Dólar"
                                        },
                                        "symbol": {
                                            "type": "string",
                                            "example": "$"
                                        },
                                        "exchange_rate": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Divisa no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Actualizar una divisa",
                "operationId": "8a397517013e49ac5bf2db204caaa130",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Libra Esterlina"
                                    },
                                    "symbol": {
                                        "type": "string",
                                        "example": "£"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Divisa actualizada exitosamente"
                    },
                    "404": {
                        "description": "Divisa no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Eliminar una divisa",
                "operationId": "c5d45f11f33d23039acf71db15a4cf31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Divisa eliminada exitosamente"
                    },
                    "404": {
                        "description": "Divisa no encontrada"
                    }
                }
            }
        },
...
    },
...
}